DeviseOS/DeviseOS/backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.111.0
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.6.1

# Database and ORM
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.11
asyncpg==0.29.0
redis==5.2.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1

# AI and ML Processing
openai==1.58.1
anthropic==0.34.1
google-generativeai==0.8.4
langchain==0.2.16
langchain-openai==0.2.8
langchain-anthropic==0.2.8
spacy==3.10.1
textblob==0.18.0
whisper==1.1.10
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.3.1

# Vector Database
pgvector==0.2.7

# Background Tasks
celery==5.4.1
kombu==5.4.1

# Object Storage
boto3==1.37.1
minio==7.3.1

# Monitoring and Observability
prometheus-client==0.21.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation-fastapi==0.45b0
sentry-sdk[fastapi]==2.17.0

# Utilities
python-dotenv==1.0.1
pytz==2024.2
python-dateutil==2.9.0
click==8.2.1
rich==14.1.0
structlog==24.4.0

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Plugin Runtime
deno==1.44.1
wasmtime==22.0.0

# GraphQL Support
strawberry-graphql[fastapi]==0.230.0

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6 